[refactor]: moved code to different file
refactor sql functions out of neo4j file and into sql functions file
